# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Monee Local Ledger API
  version: 1.2.0
  description: Unified product users with explicitly linked Google/GitHub identities.
    System admission is separate from ledger ownership/editor/viewer memberships.
    All data access is enforced by Go. Feishu is reserved for a future provider adapter.
    Local loopback HTTP only; cloud deployment is not yet supported. Cookie-authenticated
    writes require X-Monee-CSRF. Application database schema version 2.
servers:
- url: http://127.0.0.1:4173
security:
- localBearer: []
- browserSession: []
paths:
...
  "/api/v1/auth/me":
    get:
      operationId: authState
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthState"
        default:
          "$ref": "#/components/responses/Error"
      security: []
  "/api/v1/auth/start":
    post:
      operationId: startLogin
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  url:
                    type: string
                    format: uri
                  expiresIn:
                    type: integer
                    const: 600
                required:
                - id
                - url
                - expiresIn
        default:
          "$ref": "#/components/responses/Error"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum:
                  - google
                  - github
                client:
                  type: string
                  enum:
                  - web
                  - desktop
                challenge:
                  type: string
                  description: 'Required for desktop: base64url SHA-256 of a random
                    32-byte base64url verifier. Empty for web.'
                intent:
                  type: string
                  enum:
                  - login
                  - link
                  default: login
                  description: link requires an admitted initiating session (and cookie
                    CSRF), and verifies an additional identity without switching the
                    current account.
              required:
              - provider
              - client
              additionalProperties: false
      description: Creates a bounded, 10-minute login attempt. Returns 503 when this
        provider is not configured. Navigate the browser to the returned URL; no application
        session is returned here. Binding an already registered identity creates a
        pending merge that must be explicitly confirmed by the initiating session.
  "/api/v1/auth/native/poll":
    post:
      operationId: nativeLoginResult
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                    - pending
                    - complete
                    - failed
                    - linked
                    - merge_required
                  token:
                    type: string
                    description: Opaque application session only on the single successful
                      redemption. Never put it in a URL.
                required:
                - status
                - token
        default:
          "$ref": "#/components/responses/Error"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                verifier:
                  type: string
                  pattern: "^[A-Za-z0-9_-]{43}$"
              required:
              - id
              - verifier
              additionalProperties: false
      description: Proof-bound native exchange. Poll at 2-second intervals. Expired,
        mismatched or already redeemed attempts return 401. A lost successful response
        requires starting a new login.
  "/api/v1/auth/logout":
    post:
      operationId: logout
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                required:
                - ok
        default:
          "$ref": "#/components/responses/Error"
      security: []
      description: Revokes the current session. Cookie-authenticated callers must
        supply X-Monee-CSRF. An already absent session is also successful.
      parameters:
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
  "/api/v1/admin/allowlist":
    get:
      operationId: listAllowlist
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      "$ref": "#/components/schemas/AccessEntry"
                required:
                - entries
        default:
          "$ref": "#/components/responses/Error"
      description: Superadmin only. Includes enabled/disabled members and protected
        bootstrap administrators.
    post:
      operationId: addAllowlistMember
      responses:
        default:
          "$ref": "#/components/responses/Error"
        '201':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AccessEntry"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AccessSelector"
      description: Superadmin only. Creates an ordinary member. Duplicate identity
        returns 409. GitHub username lookup failure does not grant access.
      parameters:
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
  "/api/v1/admin/allowlist/{id}":
    patch:
      operationId: changeAllowlistAccess
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AccessEntry"
        default:
          "$ref": "#/components/responses/Error"
      description: Superadmin only. Protected administrators cannot be changed. Version
        conflicts return 409. Disabling blocks subsequent data requests, including
        existing sessions.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                version:
                  type: integer
                  format: int64
              required:
              - enabled
              - version
              additionalProperties: false
  "/api/v1/admin/audit":
    get:
      operationId: accessAudit
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    maxItems: 100
                    items:
                      "$ref": "#/components/schemas/AccessAudit"
                required:
                - events
        default:
          "$ref": "#/components/responses/Error"
      description: Superadmin only. Latest 100 access changes, newest first.
  "/auth/begin":
    get:
      operationId: beginBrowserLogin
      security: []
      parameters:
      - in: query
        name: ticket
        required: true
        schema:
          type: string
      responses:
        '303':
          description: Browser redirect. Invalid, expired or cancelled attempts redirect
            to a fixed result page, never an arbitrary return URL.
          headers:
            Location:
              schema:
                type: string
  "/auth/callback/{provider}":
    get:
      operationId: oauthCallback
      security: []
      description: Requires the original browser binding cookie and one-time state;
        exchanges code with PKCE and verifies the provider identity. Only GET callbacks
        are supported.
      parameters:
      - in: path
        name: provider
        required: true
        schema:
          type: string
          enum:
          - google
          - github
      - in: query
        name: state
        required: true
        schema:
          type: string
      - in: query
        name: code
        schema:
          type: string
      - in: query
        name: error
        schema:
          type: string
      responses:
        '303':
          description: Browser redirect. Invalid, expired or cancelled attempts redirect
            to a fixed result page, never an arbitrary return URL.
          headers:
            Location:
              schema:
                type: string
...
components:
  securitySchemes:
    localBearer:
      type: http
      scheme: bearer
      description: Opaque 12-hour application session obtained only by the native
        proof-bound OAuth exchange. The connection file contains no token.
    browserSession:
      type: apiKey
      in: cookie
      name: monee_session_eLaGr4oiqzKw
      description: HttpOnly SameSite=Lax session. The cookie name uses an origin hash;
        this name is for the default origin. Follow Set-Cookie when using a different
        port.
  responses:
    Error:
      description: 400 malformed JSON; 401 absent/expired session or invalid native
        proof; 403 data access denied, admin role required, CSRF or origin rejected;
        404 unknown resource; 409 identity/version conflict; 415 non-JSON write; 422
        validation; 429 too many login attempts; 503 provider unconfigured; 500 internal
        failure. Unknown routes may use plain-text 404/405.
      content:
        application/json:
          schema:
            type: object
            required:
            - code
            - message
            properties:
              code:
                type: string
              message:
                type: string
  schemas:
...
    AuthUser:
      type: object
      properties:
        provider:
          type: string
          enum:
          - google
          - github
        subject:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        allowed:
          type: boolean
        role:
          type: string
          enum:
          - ''
          - member
          - superadmin
        id:
          type: string
          description: Stable product account UUID; independent of the login provider.
      required:
      - provider
      - subject
      - username
      - email
      - displayName
      - allowed
      - role
      - id
    AuthState:
      type: object
      properties:
        user:
          oneOf:
          - "$ref": "#/components/schemas/AuthUser"
          - type: 'null'
        providers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                enum:
                - google
                - github
              enabled:
                type: boolean
            required:
            - id
            - enabled
        csrfToken:
          type: string
          description: Proof for cookie-authenticated writes; empty when no valid
            session.
      required:
      - user
      - providers
      - csrfToken
    AccessSelector:
      type: object
      properties:
        provider:
          type: string
          enum:
          - google
          - github
        kind:
          type: string
          enum:
          - subject
          - username
          - email
        value:
          type: string
        note:
          type: string
          maxLength: 200
      required:
      - provider
      - kind
      - value
      additionalProperties: false
      description: GitHub supports numeric subject or username (resolved server-side
        at add time). Google supports subject or a trusted email invitation bound
        once on verified login. Role cannot be supplied.
    AccessEntry:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        kind:
          type: string
        value:
          type: string
        subject:
          type: string
          description: Empty while a Google email invitation awaits its first verified
            login.
        note:
          type: string
        role:
          type: string
          enum:
          - member
          - superadmin
        enabled:
          type: boolean
        protected:
          type: boolean
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - provider
      - kind
      - value
      - subject
      - note
      - role
      - enabled
      - protected
      - version
      - createdAt
    AccessAudit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actor:
          type: string
        action:
          type: string
          enum:
          - add
          - bind
          - enable
          - disable
        entryId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - actor
      - action
      - entryId
      - createdAt
...